Remove CaseAccessor(s|SQL) tests
